Rename "index" to "Summary", put routing structure into index
